test: split test_worktree_activity_probe under the 1500-line guardrail
The module had grown to 1722 lines and tripped the first-party line-limit
guardrail in test_core_decomposition_maintainability. Split it into three
part modules under test_worktree_activity_probe_parts/, following the
existing *_parts/ convention: shared helpers move to helpers.py, the
worktree fixture to conftest.py, and the tests split at thematic seams —
tree changes and linked-git-dir watching (part 001), walk uncertainty and
priming (part 002), scan threads and the worker-wide ceiling (part 003).

Each part is well under 1350 lines so later review rounds have headroom.
Test bodies are byte-identical to the original; no behaviour change.
